pyproject.toml
[tool.poetry]
name = "labelprinterkit"
version = "0.4.0"
description = "A library for creating and printing labels for Brother P-Touch devices"
authors = ["Adrian Tschira <packages@notafile.com>", "Benedikt Neuffer <benedikt.neuffer@kit.edu>"]
license = "Apache License, Version 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.5.0"
pyusb = "^1.2.1"
packbits = "^0.6"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
